requirements.txt
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow>=10.2.0,<11.0.0
surya-ocr==0.17.1
pyzbar>=0.1.9
opencv-python-headless>=4.8.0
pydantic-settings>=2.0.0
platformdirs>=4.0.0
ftfy>=6.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langdetect==1.0.9
numpy>=1.24.0,<2.0.0
transformers>=4.40.0,<5.0.0
sentencepiece>=0.2.0
sacremoses>=0.1.1
torch>=2.7.0,<2.11.0
